refactor: remove few barrel exports
