+++ apps/web/app/(all)/[workspaceSlug]/(projects)/projects/(detail)/[projectId]/intake/layout.tsx
@@ -3,7 +3,7 @@
 // components
 import { AppHeader } from "@/components/core/app-header";
 import { ContentWrapper } from "@/components/core/content-wrapper";
-import { ProjectInboxHeader } from "@/plane-web/components/projects/settings/intake";
+import { ProjectInboxHeader } from "@/plane-web/components/projects/settings/intake/header";
 
 export default function ProjectInboxIssuesLayout({ children }: { children: React.ReactNode }) {
   return (
