--- apps/web/ce/components/common/index.ts
@@ -1,2 +1,0 @@
-export * from "./subscription";
-export * from "./extended-app-header";
