--- apps/web/ce/components/common/subscription/index.ts
@@ -1,1 +1,0 @@
-export * from "./subscription-pill";
