--- apps/web/ce/components/de-dupe/index.ts
@@ -1,4 +1,0 @@
-export * from "./de-dupe-button";
-export * from "./duplicate-modal";
-export * from "./duplicate-popover";
-export * from "./issue-block";
