--- apps/web/ce/components/de-dupe/issue-block/index.ts
@@ -1,1 +1,0 @@
-export * from "./button-label";
