--- apps/web/ce/components/epics/index.ts
@@ -1,1 +1,0 @@
-export * from "./epic-modal";
