--- apps/web/ce/components/projects/navigation/index.ts
@@ -1,1 +1,0 @@
-export * from "./helper";
