--- apps/web/ce/components/projects/settings/intake/index.ts
@@ -1,1 +1,0 @@
-export * from "./header";
