--- apps/web/ce/components/projects/teamspaces/index.ts
@@ -1,1 +1,0 @@
-export * from "./teamspace-list";
