+++ apps/web/core/components/settings/sidebar/header.tsx
@@ -6,7 +6,7 @@ import { WorkspaceLogo } from "@/components/workspace/logo";
 // hooks
 import { useWorkspace } from "@/hooks/store/use-workspace";
 // plane web imports
-import { SubscriptionPill } from "@/plane-web/components/common";
+import { SubscriptionPill } from "@/plane-web/components/common/subscription/subscription-pill";
 
 export const SettingsSidebarHeader = observer((props: { customHeader?: React.ReactNode }) => {
   const { customHeader } = props;
