--- apps/web/ee/components/de-dupe/index.ts
@@ -1,1 +1,0 @@
-export * from "ce/components/de-dupe";
