--- apps/web/ee/components/epics/index.ts
@@ -1,1 +1,0 @@
-export * from "ce/components/epics";
